Split index into two files to handle function-options in a linter friendly way
